Split cloud_trace.[h|cc] into sampler/aggregator/cloud_trace.
