+++ src/api_manager/cloud_trace/cloud_trace.cc
@@ -14,7 +14,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////////
 //
-#include "cloud_trace.h"
+#include "src/api_manager/cloud_trace/cloud_trace.h"
 
 #include <cctype>
 #include <chrono>
@@ -27,11 +27,8 @@
 #include "google/protobuf/timestamp.pb.h"
 #include "include/api_manager/utils/status.h"
 #include "include/api_manager/utils/version.h"
-#include "src/api_manager/utils/marshalling.h"
 
-using google::api_manager::utils::Status;
 using google::devtools::cloudtrace::v1::Trace;
-using google::devtools::cloudtrace::v1::Traces;
 using google::devtools::cloudtrace::v1::TraceSpan;
 using google::devtools::cloudtrace::v1::TraceSpan_SpanKind;
 using google::protobuf::Timestamp;
@@ -41,7 +38,6 @@ namespace api_manager {
 namespace cloud_trace {
 namespace {
 
-const char kCloudTraceService[] = "/google.devtools.cloudtrace.v1.TraceService";
 // Cloud Trace agent label key
 const char kCloudTraceAgentKey[] = "trace.cloud.google.com/agent";
 // Cloud Trace agent label value
@@ -92,117 +88,6 @@ void GetTraceFromGRpcTraceContextHeader(const std::string &raw_trace_context,
                                         const std::string &root_span_name,
                                         Trace **trace, std::string *options);
 }  // namespace
-
-Sampler::Sampler(double qps) {
-  if (qps == 0.0) {
-    is_disabled_ = true;
-  } else {
-    duration_ = 1.0 / qps;
-    is_disabled_ = false;
-  }
-}
-
-bool Sampler::On() {
-  if (is_disabled_) {
-    return false;
-  }
-  auto now = std::chrono::system_clock::now();
-  std::chrono::duration<double> diff = now - previous_;
-  if (diff.count() > duration_) {
-    previous_ = now;
-    return true;
-  } else {
-    return false;
-  }
-};
-
-void Sampler::Refresh() {
-  if (is_disabled_) {
-    return;
-  }
-  previous_ = std::chrono::system_clock::now();
-}
-
-Aggregator::Aggregator(auth::ServiceAccountToken *sa_token,
-                       const std::string &cloud_trace_address,
-                       int aggregate_time_millisec, int cache_max_size,
-                       double minimum_qps, ApiManagerEnvInterface *env)
-    : sa_token_(sa_token),
-      cloud_trace_address_(cloud_trace_address),
-      aggregate_time_millisec_(aggregate_time_millisec),
-      cache_max_size_(cache_max_size),
-      traces_(new Traces),
-      env_(env),
-      sampler_(minimum_qps) {
-  sa_token_->SetAudience(auth::ServiceAccountToken::JWT_TOKEN_FOR_CLOUD_TRACING,
-                         cloud_trace_address_ + kCloudTraceService);
-}
-
-void Aggregator::Init() {
-  if (aggregate_time_millisec_ == 0) {
-    return;
-  }
-  timer_ = env_->StartPeriodicTimer(
-      std::chrono::milliseconds(aggregate_time_millisec_),
-      [this]() { SendAndClearTraces(); });
-}
-
-Aggregator::~Aggregator() {
-  if (timer_) {
-    timer_->Stop();
-  }
-}
-
-void Aggregator::SendAndClearTraces() {
-  if (traces_->traces_size() == 0 || project_id_.empty()) {
-    env_->LogDebug(
-        "Not sending request to CloudTrace: no traces or "
-        "project_id is empty.");
-    traces_->clear_traces();
-    return;
-  }
-
-  // Add project id into each trace object.
-  for (int i = 0; i < traces_->traces_size(); ++i) {
-    traces_->mutable_traces(i)->set_project_id(project_id_);
-  }
-
-  std::unique_ptr<HTTPRequest> http_request(new HTTPRequest(
-      [this](Status status, std::map<std::string, std::string> &&,
-             std::string &&body) {
-        if (status.code() < 0) {
-          env_->LogError("Trace Request Failed." + status.ToString());
-        } else {
-          env_->LogDebug("Trace Response: " + status.ToString() + "\n" + body);
-        }
-      }));
-
-  std::string url =
-      cloud_trace_address_ + "/v1/projects/" + project_id_ + "/traces";
-
-  std::string request_body;
-
-  ProtoToJson(*traces_, &request_body, utils::DEFAULT);
-  traces_->clear_traces();
-  env_->LogDebug("Sending request to Cloud Trace.");
-  env_->LogDebug(request_body);
-
-  http_request->set_url(url)
-      .set_method("PATCH")
-      .set_auth_token(sa_token_->GetAuthToken(
-          auth::ServiceAccountToken::JWT_TOKEN_FOR_CLOUD_TRACING))
-      .set_header("Content-Type", "application/json")
-      .set_body(request_body);
-
-  env_->RunHTTPRequest(std::move(http_request));
-}
-
-void Aggregator::AppendTrace(google::devtools::cloudtrace::v1::Trace *trace) {
-  traces_->mutable_traces()->AddAllocated(trace);
-  if (traces_->traces_size() > cache_max_size_) {
-    SendAndClearTraces();
-  }
-}
 
 CloudTrace::CloudTrace(Trace *trace, const std::string &options,
                        HeaderType header_type)
@@ -284,11 +169,11 @@ CloudTraceSpan::~CloudTraceSpan() {
     return;
   }
   GetNow(trace_span_->mutable_end_time());
-  for (unsigned int i = 0; i < messages.size(); ++i) {
+  for (unsigned int i = 0; i < messages_.size(); ++i) {
     std::stringstream stream;
     stream << std::setfill('0') << std::setw(3) << i;
     std::string sequence = stream.str();
-    trace_span_->mutable_labels()->insert({sequence, messages[i]});
+    trace_span_->mutable_labels()->insert({sequence, messages_[i]});
   }
 }
 
@@ -297,7 +182,7 @@ void CloudTraceSpan::Write(const std::string &msg) {
     // Trace is disabled.
     return;
   }
-  messages.push_back(msg);
+  messages_.push_back(msg);
 }
 
 CloudTrace *CreateCloudTrace(const std::string &trace_context,
