+++ src/api_manager/cloud_trace/cloud_trace.h
@@ -21,9 +21,8 @@
 
 #include "google/devtools/cloudtrace/v1/trace.pb.h"
 #include "google/protobuf/map.h"
-#include "include/api_manager/env_interface.h"
-#include "include/api_manager/periodic_timer.h"
-#include "src/api_manager/auth/service_account_token.h"
+#include "src/api_manager/cloud_trace/aggregator.h"
+#include "src/api_manager/cloud_trace/sampler.h"
 
 namespace google {
 namespace api_manager {
@@ -32,86 +31,6 @@ namespace cloud_trace {
 enum HeaderType {
   CLOUD_TRACE_CONTEXT = 1,
   GRPC_TRACE_CONTEXT = 2,
-};
-
-// A helper class to determine if trace should be enabled for a request.
-// A Sampler instance is put into the Aggregator class.
-// Trace is triggered if the time interval between the request time and the
-// previous trace enabled request is bigger than a threshold.
-// The threshold is calculated from the qps.
-class Sampler {
- public:
-  Sampler(double qps);
-
-  // Returns whether trace should be turned on for this request.
-  bool On();
-
-  // Refresh the previous timestamp to the current time.
-  void Refresh();
-
- private:
-  bool is_disabled_;
-  std::chrono::time_point<std::chrono::system_clock> previous_;
-  double duration_;
-};
-
-// TODO: The Aggregator class is not thread safe.
-// TODO: simplify class naming in this file.
-// Stores cloud trace configurations shared within the job. There should be
-// only one such instance. The instance is put in service_context.
-class Aggregator final {
- public:
-  Aggregator(auth::ServiceAccountToken *sa_token,
-             const std::string &cloud_trace_address,
-             int aggregate_time_millisec, int cache_max_size,
-             double minimum_qps, ApiManagerEnvInterface *env);
-
-  ~Aggregator();
-
-  // Initializes the aggregator by setting up a periodic timer. At each timer
-  // invocation traces aggregated are sent to Cloud Trace API
-  void Init();
-
-  // Flush traces cached and clear the traces_ proto.
-  void SendAndClearTraces();
-
-  // Append a Trace to traces_, the appended trace may not be sent at the time
-  // of this function call.
-  void AppendTrace(google::devtools::cloudtrace::v1::Trace *trace);
-
-  // Sets the producer project id
-  void SetProjectId(const std::string &project_id) { project_id_ = project_id; }
-
-  // Get the sampler.
-  Sampler &sampler() { return sampler_; }
-
- private:
-  // ServiceAccountToken object to get auth tokens for Cloud Trace API.
-  auth::ServiceAccountToken *sa_token_;
-
-  // Address for Cloud Trace API
-  std::string cloud_trace_address_;
-
-  // The maximum time to hold a trace before sent to Cloud Trace.
-  int aggregate_time_millisec_;
-
-  // The maximum number of traces that can be cached.
-  int cache_max_size_;
-
-  // Traces protobuf to hold a list of Trace obejcts.
-  std::unique_ptr<google::devtools::cloudtrace::v1::Traces> traces_;
-
-  // The producer project id.
-  std::string project_id_;
-
-  // ApiManager Env used to set up periodic timer.
-  ApiManagerEnvInterface *env_;
-
-  // Timer to trigger flush trace.
-  std::unique_ptr<google::api_manager::PeriodicTimer> timer_;
-
-  // Sampler object to help determine if trace should be enabled for a request.
-  Sampler sampler_;
 };
 
 // Stores traces and metadata for one request. The instance of this class is
@@ -191,7 +110,7 @@ class CloudTraceSpan {
                             protobuf::uint64 parent_span_id);
   CloudTrace *cloud_trace_;
   google::devtools::cloudtrace::v1::TraceSpan *trace_span_;
-  std::vector<std::string> messages;
+  std::vector<std::string> messages_;
 };
 
 // Parses the trace_context and determines if cloud trace should
